Simplify pointer encoding
The obfuscation of pointers depends on having a part of the pointer
unused (i.e. always 0) to embed a checksum for pointer verification.

Since this is true for Linux on the most common architectures, instead
of returning an error that needs to be checked, just abort in case one
such pointer is found. This probably means a corrupted pointer which
would also cause a crash elsewhere.

Signed-off-by: Xavi Hernandez <xhernandez@gmail.com>