Implement CWD support
We need to filter any attempt to access parent directories of the root
inode of the mount. Since the underlying instance of the ceph mount has
been connected to "/", the libcephfs library is not aware of the
effective root mount point we are using. To avoid that the user can
access unwanted directories by using relative paths like ".." or
specially built symbolic links, we need to control all path walking
operations in the proxy itself.

As part of this change, we also implement the operations chdir() and
getcwd() to use internal paths independent of libcephfs.

Signed-off-by: Xavi Hernandez <xhernandez@gmail.com>